# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is official backend documentation for the fngc software",
        "title": "FNGC Backend Service",
        "contact": {
            "name": "Endy Apinageri",
            "email": "apinaendy@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "petstore.swagger.io",
    "basePath": "/v1",
    "paths": {
...
        "/student/curriculum/": {
            "get": {
                "description": "Gets the list of all curricul courses  being taken by the student",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Student"
                ],
                "summary": "Handles getting student curriculum",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/models.StudentCurriculum"
                        }
                    },
                    "400": {
                        "description": "Check Response Message",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseBody"
                        }
                    }
                }
            }
        },
        "/student/examination/": {
            "post": {
                "description": "Accept JSON data for examination objects and returns valid response",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Student"
                ],
                "summary": "Handle apply for student examination",
                "parameters": [
                    {
                        "description": "The Student Examination Data",
                        "name": "ExaminationPreparationData",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ExamPreparation"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/models.ExamPreparation"
                        }
                    },
                    "400": {
                        "description": "Check Response Message",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseBody"
                        }
                    }
                }
            }
        },
        "/student/tutor/request": {
            "post": {
                "description": "Accept JSON data for new tutor request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Student"
                ],
                "summary": "Handle student request for new tutor",
                "parameters": [
                    {
                        "description": "The Student Examination Data",
                        "name": "RequestData",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.RequestTutor"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/models.RequestTutor"
                        }
                    },
                    "400": {
                        "description": "Check Response Message",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseBody"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.ExamPreparation": {
            "type": "object",
            "properties": {
                "class_type": {
                    "type": "string"
                },
                "exam_cost": {
                    "type": "string"
                },
                "exam_id": {
                    "type": "string"
                },
                "exam_profile": {
                    "type": "string"
                },
                "ezam_status": {
                    "type": "string"
                },
                "personal_tutor": {
                    "type": "boolean"
                },
                "study_pack": {
                    "type": "boolean"
                },
                "training_duration": {
                    "type": "string"
                },
                "training_type": {
                    "type": "string"
                },
                "tutor_type": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "models.RequestTutor": {
            "type": "object",
            "properties": {
                "exam_id": {
                    "type": "string"
                },
                "exam_type": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "student_id": {
                    "type": "string"
                }
            }
        },
...
        "models.ResponseBody": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "object"
                },
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "models.StudentCurriculum": {
            "type": "object",
            "properties": {
                "class": {
                    "type": "string"
                },
                "class_type": {
                    "type": "string"
                },
                "subjject": {
                    "type": "string"
                }
            }
        },
...
    }
}